Moved rosetta hit to own file and slightly refactored run method
